requirements.txt
arxiv>=2.1.0
openai>=1.12.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0